feat: add a simple search box on problem set page
Add a search box on problem set page. Only uses a simple algorithm, you can't use it like Google. You can search one or some of the words in the title.

NOTICE: The style of the sarch box may have some problems. when you turn the window into mobile mode, the search box will not be at the center of the page. Hope someone can make it better.