fix(install/bundle): fix the wrong generation of our-root password and length
The last commit changed the installation script a lot, but made some mistakes on the genration of config.
It gens the $_svn_ourroot_password_ to $_svn_certroot_password_. Just fix it.
According to the version from @vfleaking, the judger_socket password and the our-root password is 32 chars long. Strengthen it.
There is also a *** space char on the end of a line. Kick it out.