fix(uoj/1): change server time's timezone that always set to UTC
Since PHP 5.1.0 (when the date/time functions were rewritten), we need to specify the default timezone.
Otherwise the timezone will always set to UTC that may cause problem or cause confusion.
We set the default timezone to Asia/Shanghai. If you are not at the area of this timezone, you can change it manually.