style(judge_client/1/uoj_judger): change "Judgment" spelling
Okay, in truth it is not really a bug.
We just change its name in php because function lacks.
For judger it will work correctly and will not crash during running the program.
But, to avoid inconsistence we use "Judgement" for all purpose.

"Judgment" or "Judgement"? They are both right.
You ask me which is better? I don't know.
I make the choice only because the word with more "e" is used almost everywhere.
Used more frequently and I don't want to edit a lot of files.