fix(blog-preview.php): click readmore will lead to 404 page
On Blog Preview page when someone set readmore flag will make rest of the content invisible.
Instead there will be a button for users to click and get the full article.
Because of the wrongly put link, it will make the last slash missing.
Put the blog id back in the right place to get a right link.