feat(web):make download easier
Show the download_file on the problem page if needed and show the data_file if the user is SuperUser.
What's more, hide the url in the problem_manege page if it leads to 404 page.
In passing, support that SuperUser can download the data_file by url: /download?type=data&id=problem_id